Fixing a subset of the issues raised by Matt
(1) The UnquotedString token matches the empty string.  I was wondering whether it's possible to remove this token altogether (see comments below)

RESPONSE:
TODO

(2) Could you move the definition of Alpha-Char further up in the spec (after 2.2.1?) as I got to 2.2.2 and was wondering what an Alpha-Char was - at first, I didn't see that it appear later on.

RESPONSE:
OK, now 2.2.0

(3) It seems like the ordering of id token specification is important.  i.e.  Canonical-Prefixed then NonCanonical-Prefixed then Un-prefixed then URL-as-ID.  Tokens should be matched in that order.  Otherwise, some could be matched as others.

RESPONSE:

well it shouldn't matter for non-canonical vs canonical, as it's rule is:

NonCanonical-Prefixed-ID ::= ( Any-IDPrefix ':' Any-LocalID ) - Canonical-Prefixed-ID

and unprefixed IDs can't have ":"s:

Unprefixed-ID ::= { ( NonWsChar - ':' ) }



(4) The token spec for NonCanonical-Prefixed-ID ends up with it having two colon characters (Any-IDPrefix has a terminal colon in it and then NonCanonical-Prefixed-ID has a terminal colon in it).

RESPONSE:
TODO

(5) I'm not totally sure, but it looks like all ID tokens could be matched by the UnquotedString token (actually all tokens - ordering really matters).  Just looking at where this token is used, it might be good if it could be removed altogether.

RESPONSE:

That's fine - there is deliberately no restriction on UnquotedString - it's used in places where any arbitrary text is allowed - for example, a name: clause.

(6) UnreservedToken is not defined anywhere.  (I found this after looking for uses of UnquotedString)

RESPONSE:

Yes, not sure what to do with this. We've historically allowed arbitrary tag-value pairs, with the recommendation that parses roundtrip them. This means that an obof1.0 parser could roundtrip things like the "intersection_of" tags introduced in obof1.2 which turned out to be pretty useful.

However, there will never be an obof higher than 1.4, so we could simply eliminate this production, and get more error checking into the bargain.